# services/order/doc/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyApp Order Processing Service
  description: |-
    This is API documentation of my order processing server based on the OpenAPI 3.0 specification.
    termsOfService: http://swagger.io/terms/
  contact:
    email: xyz@my-orderproc-app.io
  license:
    name: License Not Yet (TODO)
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
  - url: https://doc.site.not.ready/api/v1.1.0
...
paths:
...
  /order/{order-id}:
    get:
      summary: Get detail of an order
      description: Get some properties of an existing order.
      tags:
        - order
      parameters:
        - name: order-id
          in: path
          description: Identifier for the order
          required: true
          schema:
            type: string
        - name: shipping
          in: query
          description: 'load shipping detail'
          schema:
            type: boolean
        - name: billing
          in: query
          description: 'load billing detail'
          schema:
            type: boolean
        - name: lines
          in: query
          description: load order lines grouped by different states (pending, reserved, paid, shipping ...etc).
          schema:
            type: boolean
      responses:
        '200':
          description: the resulting data from successful read operation will vary based on specific query parameters you specified in the URL path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderInfoRead'
        '404':
          description: Order not found
        '503':
          description: Service unavailable
      security:
        - BearerAuth: []
    patch:
      summary: Update detail of an order
      description: Update properties of an existing order, except chosen product items
      tags:
        - order
      parameters:
        - name: order-id
          in: path
          description: Identifier for the order
          required: true
          schema:
            type: string
        - name: shipping
          in: query
          description: 'enable update on shipping'
          schema:
            type: boolean
        - name: billing
          in: query
          description: 'enable update on billing'
          schema:
            type: boolean
      requestBody:
        required: true
        description: the properties to edit should correspond to specific query parameters you specified in the URL path
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequestEdit'
      responses:
        '200':
          description: Successful update
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEditErrorResponse'
        '503':
          description: Service unavailable
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    OrderInfoRead:
      type: object
      properties:
        reserved_lines:
          type: array
          items:
            anyOf:
            - $ref: '#/components/schemas/PayOrderLineObject'
        paid_lines:
          type: array
          items:
            anyOf:
            - $ref: '#/components/schemas/PayOrderLineObject'
            - $ref: '#/components/schemas/PayOrderLineObject'
        shipping_lines:
          type: array
          items:
            oneOf:
            - $ref: '#/components/schemas/OrderLineObject'
        completed_lines:
          type: array
          items:
            oneOf:
            - $ref: '#/components/schemas/OrderLineObject'
          minItems: 0
        shipping:
          anyOf:
          - $ref: '#/components/schemas/ShippingValid'
          properties:
            shipments:
              type: array
              items:
                type: string
              description: list of shipment IDs associated with the order
              example: ['8x9e8r2tE', 'JiHuIIiKKUkm']
        billing:
          anyOf:
          - $ref: '#/components/schemas/BillingValid'
          properties:
            payments:
              type: array
              items:
                type: string
              description: list of payment IDs associated with the order
              example: ['9j8cj2r', '01k3kl1iK']
    OrderRequestEdit:
      type: object
      properties:
        shipping:
          $ref: '#/components/schemas/ShippingValid'
        billing:
          $ref: '#/components/schemas/BillingValid'
    BaseOrderLineObject:
      type: object
      properties:
        product_type:
          type: integer
          format: uint8
          example: 1
        product_id:
          type: integer
          format: uint64
          example: 270118
        quantity:
          type: integer
          format: uint32
          example: 5
    PayAmountObject:
      type: object
      properties:
        unit:
          type: integer
          format: uint32
          example: 9
        total:
          type: integer
          format: uint32
          description: in case the seller provides discount or couple in the future.
          example: 44
    PayOrderLineObject:
      anyOf:
        - $ref: '#/components/schemas/BaseOrderLineObject'
      properties:
        seller_id:
          type: integer
          format: uint32
          example: 986543
        amount:
          $ref: '#/components/schemas/PayAmountObject'
      required:
        - seller_id
        - product_type
        - product_id
        - quantity
        - amount
...
    OrderLineObject:
      anyOf:
        - $ref: '#/components/schemas/BaseOrderLineObject'
      properties:
        seller_id:
          type: integer
          format: uint32
          example: 18830
      required:
        - seller_id
        - product_type
        - product_id
        - quantity
...
    BillingValid:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/AddressValid'
        contact:
          $ref: '#/components/schemas/ContactValid'
      required:
        - contact
    BillingErrDetail:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/AddressErrDetail'
        contact:
          $ref: '#/components/schemas/ContactErrDetail'
        nonfield:
          type: string
          enum: [contact-required]
    ShippingValid:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/AddressValid'
        contact:
          $ref: '#/components/schemas/ContactValid'
        option:
          $ref: '#/components/schemas/ShippingOptionValid'
      required:
        - contact
        - option
    ShippingErrDetail:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/AddressErrDetail'
        contact:
          $ref: '#/components/schemas/ContactErrDetail'
        option:
          $ref: '#/components/schemas/ShippingOptionErrDetail'
        non-field:
          type: string
          enum: [contact-required, option-required]
    AddressValid:
      type: object
      properties:
        country:
          type: string
          enum: [TW, IN, TH, ID, US]
        region:
          type: string
          example: Main island
        city:
          type: string
          example: New Taipei City
        distinct:
          type: string
          example: xyz678
        street_name:
          type: string
          example: QWETY st.
        detail:
          type: string
          example: 1-24-4 zabc building 2351
      required:
        - country
        - region
        - city
        - distinct
        - detail
    AddressErrDetail:
      type: object
      properties:
        country:
          type: string
          enum: [empty, invalid-char, not-exist, not-support]
          example: not-support
        region:
          type: string
          enum: [empty, invalid-char, not-exist, not-support]
          example: not-exist
        city:
          type: string
          enum: [empty, invalid-char, not-exist]
          example: invalid-char
        distinct:
          type: string
          enum: [empty, invalid-char]
        street_name:
          type: string
          enum: [empty, invalid-char]
        detail:
          type: string
          enum: [empty, invalid-char]
    ShippingMethodObject:
      type: string
      enum: [FedEx, UPS]
    ShippingOptionValid:
      type: array
      items:
        type: object
        properties:
          seller_id:
            type: integer
            format: uint32
          method:
            $ref: '#/components/schemas/ShippingMethodObject'
        required:
          - method
          - seller_id
      example:
        - method: UPS
          seller_id: 27037
        - method: FedEx
          seller_id: 18830
    ShippingOptionErrDetail:
      type: array
      items:
        type: object
        properties:
          seller_id:
            type: string
            enum: [empty, not-exist, not-support]
            description: the error message `not-support` is present when user assigns an existing seller / store unrelated to the given order.
          method:
            type: string
            enum: [empty, not-support]
      example:
        - method: not-support
          seller_id: not-exist
        - method: empty
          seller_id: empty
    ContactValid:
      type: object
      properties:
        first_name:
          type: string
          example: Jon
        last_name:
          type: string
          example: Snow
        emails:
          type: array
          items:
            type: string
          example: ['tobi@ninjacat.io', 'mindful@iowalab.idv', 'move@aheadof.light']
          minItems: 1
        phones:
          type: array
          items:
            anyOf:
            - $ref: '#/components/schemas/PhoneNumberValid'
            - $ref: '#/components/schemas/PhoneNumberValid'
          minItems: 1
      required:
        - first_name
        - last_name
        - emails
        - phones
    ContactErrDetail:
      type: object
      properties:
        first_name:
          type: string
          enum: [empty, invalid-char]
        last_name:
          type: string
          enum: [empty, invalid-char]
        emails:
          type: array
          description: error detail is located at corresponding index of the array element, `null` means the email has valid format.
          items:
            type: string
            enum: [invalid-code, invalid-char]
          example: [null, 'invalid-code', 'invalid-char']
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneNumberErrDetail'
        non-field:
            type: string
            enum: [email-missing, phone-missing]
    PhoneNumberValid:
      type: object
      properties:
        nation:
          type: integer
          format: uint16
          example: 886
        number:
          type: string
          example: '09138411'
      required:
        - nation
        - number
    PhoneNumberErrDetail:
      type: object
      properties:
        nation:
          type: string
          enum: [invalid-code]
        number:
          type: string
          enum: [invalid-char, empty]
...
    OrderEditErrorResponse:
      type: object
      properties:
        shipping:
          $ref: '#/components/schemas/ShippingErrDetail'
        billing:
          $ref: '#/components/schemas/BillingErrDetail'
        non-field:
          type: string
          enum: [billing-locked, shipping-locked, order-archived]
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT